move bank functions to bank folder
