+++ app/routes/__test__/api.banks.list.test.ts
@@ -1,6 +1,6 @@
 import { type LoaderFunctionArgs } from "react-router";
 import { Result } from "true-myth";
-import { type BankData } from "~/components/formElements/inputs/iban/bankNameFromIBAN";
+import { type BankData } from "~/services/bank/bankNameFromIBAN";
 import { loader } from "~/routes/api.banks.list";
 import { validateCsrfSessionFormless } from "~/services/security/csrf/validatedSession.server";
